Move path_to_str to another file
